requirements.txt
asgiref==3.2.3
astroid==2.3.3
attrs==19.3.0
Automat==20.2.0
CacheControl==0.12.6
cachy==0.3.0
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
cleo==0.7.6
click==7.1.2
clikit==0.4.3
colorama==0.4.3
constantly==15.1.0
cryptography==2.9.2
cssselect==1.1.0
databases==0.3.2
Django==2.2
fastapi==0.58.1
Flask==1.1.2
h11==0.9.0
html5lib==1.1
hyperlink==19.0.0
idna==2.10
importlib-metadata==1.1.3
incremental==17.5.0
isort==4.3.21
itemadapter==0.1.0
itsdangerous==1.1.0
Jinja2==2.11.2
jsonschema==3.2.0
keyring==20.0.1
lazy-object-proxy==1.4.3
lockfile==0.12.2
lxml==4.5.2
MarkupSafe==1.1.1
mccabe==0.6.1
msgpack==1.0.0
mysqlclient==1.4.6
parsel==1.6.0
pastel==0.2.0
pexpect==4.8.0
pkginfo==1.5.0.1
poetry==1.0.9
Protego==0.1.16
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pydantic==1.5.1
PyDispatcher==2.0.5
PyHamcrest==2.0.2
PyJWT==1.7.1
pylev==1.3.0
pylint==2.4.4
PyMySQL==0.9.3
pyOpenSSL==19.1.0
pyparsing==2.4.7
pyrsistent==0.14.11
python-multipart==0.0.5
pytz==2019.3
pywin32-ctypes==0.2.0
queuelib==1.5.0
requests==2.24.0
requests-toolbelt==0.8.0
Scrapy==2.2.0
service-identity==18.1.0
shellingham==1.3.2
six==1.13.0
SQLAlchemy==1.3.18
sqlparse==0.3.0
starlette==0.13.4
tomlkit==0.5.11
Twisted==20.3.0
typed-ast==1.4.0
urllib3==1.25.9
uvicorn==0.11.5
virtualenv==16.4.3
w3lib==1.22.0
webencodings==0.5.1
websockets==8.1
Werkzeug==1.0.1
wrapt==1.11.2
yapf==0.29.0
zipp==3.1.0
zope.interface==5.1.0
